fix: Clean up spacing, indentation, expand on docstrings, etc.
