+++ app/src/main/java/no/usn/mob3000/data/repository/social/ProfileEditRepository.kt
@@ -6,7 +6,6 @@ import no.usn.mob3000.data.source.remote.auth.UserDataSource
 import no.usn.mob3000.data.source.remote.social.ProfileUserDataSource
 import no.usn.mob3000.domain.repository.social.IProfileEditRepository
 /**
- *
  * This repository orchestrates the updating of the users profile. Currently there is only one implementation of
  * updating a table under profile, thus [updateProfile] being the sole function here
  *
@@ -17,7 +16,7 @@ import no.usn.mob3000.domain.repository.social.IProfileEditRepository
 class ProfileEditRepository (
     private val profileEditDataSource: ProfileUserDataSource = ProfileUserDataSource(),
     private val userDataSource: UserDataSource = UserDataSource()
-): IProfileEditRepository{
+) : IProfileEditRepository {
     /**
      *  Updates the users profile. As long as the value of the user exist, we try to update. Else throw exception
      *
@@ -37,21 +36,22 @@ class ProfileEditRepository (
         nationality: String,
         profileVisibility: Boolean,
         friendsVisibility: Boolean
-    ) : Result<Unit>{
+    ): Result<Unit> {
         val originalUser = userDataSource.fetchUserById(userid)
-        return if( originalUser != null ) {
+
+        return if (originalUser != null) {
             val updateProfileDto = ProfileDto(
                 userId = userid,
-                    updatedAt = Clock.System.now(),
-                    displayName = displayName,
-                    avatarUrl = avatarUrl,
-                    aboutMe = aboutMe,
-                    nationality = nationality,
-                    profileVisibility = profileVisibility,
-                    friendsVisibility = friendsVisibility
+                updatedAt = Clock.System.now(),
+                displayName = displayName,
+                avatarUrl = avatarUrl,
+                aboutMe = aboutMe,
+                nationality = nationality,
+                profileVisibility = profileVisibility,
+                friendsVisibility = friendsVisibility
             )
             profileEditDataSource.updateProfile(userid,updateProfileDto)
-        } else{
+        } else {
             Result.failure(Exception("Original User data not found"))
         }
     }
