+++ app/src/main/java/no/usn/mob3000/data/source/remote/social/FriendRequestDataSource.kt
@@ -8,8 +8,8 @@ import kotlinx.coroutines.withContext
 import no.usn.mob3000.data.model.social.FriendRequestsDto
 import no.usn.mob3000.data.model.social.FriendsDto
 import no.usn.mob3000.data.network.SupabaseClientWrapper
+
 /**
- *
  * A data source class responsible for handling friend request-related calls via Supabase.
  *
  * @param supabaseClient The Supabase client for interacting with the backend.
@@ -17,10 +17,10 @@ import no.usn.mob3000.data.network.SupabaseClientWrapper
  * @author Husseinabdulameer11, 258030
  * @created 2024-11-13
  */
-
- class FriendRequestDataSource (private val supabaseClient: SupabaseClient = SupabaseClientWrapper.getClient()){
+class FriendRequestDataSource(
+    private val supabaseClient: SupabaseClient = SupabaseClientWrapper.getClient()
+) {
     /**
-     *
      * This function retrieves a list of all friend requests where the user is the "to_user".
      *
      * @param userId The ID of the user to fetch friend requests for.
@@ -31,13 +31,14 @@ import no.usn.mob3000.data.network.SupabaseClientWrapper
     suspend fun fetchFriendRequests(userId: String): List<FriendRequestsDto> = withContext(Dispatchers.IO) {
         supabaseClient
             .from("friend_requests")
-            .select()
-            { filter { eq("to_user", userId) } }
-            .decodeList()
+            .select() {
+                filter {
+                    eq("to_user", userId)
+                }
+            }.decodeList()
     }
 
     /**
-     *
      * This function retrieves a list of all friend requests where the user is the "by_user" or "to_user".
      *
      * @param userId The ID of the user to fetch friend requests for.
@@ -55,11 +56,10 @@ import no.usn.mob3000.data.network.SupabaseClientWrapper
                         eq("to_user", userId)
                     }
                 }
-            }
-            .decodeList()
+            }.decodeList()
     }
+
     /**
-     *
      * This function retrieves a single friend request based on the given ID, used for getting a specific request
      * for further actions.
      *
@@ -71,14 +71,14 @@ import no.usn.mob3000.data.network.SupabaseClientWrapper
     suspend fun fetchFriendRequestById(friendRequestId: String): FriendRequestsDto? = withContext(Dispatchers.IO) {
         supabaseClient
             .from("friend_requests")
-            .select()
-            {
-                filter { eq("id", friendRequestId) }
-            }
-            .decodeSingleOrNull()
+            .select {
+                filter {
+                    eq("id", friendRequestId)
+                }
+            }.decodeSingleOrNull()
     }
+
     /**
-     *
      * Removes a a friend request based on the requestId.  Used for removing a friend request after it has been either accepted or cancelled.
      *
      * @param friendRequestId The ID of the friend request to delete.
@@ -88,12 +88,16 @@ import no.usn.mob3000.data.network.SupabaseClientWrapper
     suspend fun deleteFriendRequestById(friendRequestId: String): PostgrestResult = withContext(Dispatchers.IO) {
         supabaseClient
             .from("friend_requests")
-            .delete { filter { eq("id", friendRequestId) } }
+            .delete {
+                filter {
+                    eq("id", friendRequestId)
+                }
+            }
     }
+
     /**
-     *
-     * Used when the user sends a friend request. Creates a new row in the friend_requests table, where the "requested" friend have the option of accepting or
-     * declining the invitation.
+     * Used when the user sends a friend request. Creates a new row in the friend_requests table,
+     * where the "requested" friend have the option of accepting or declining the invitation.
      *
      * @param friendRequestItem An instance of [FriendRequestsDto] containing the data for the new friend request.
      * @author Husseinabdulameer11
@@ -101,14 +105,17 @@ import no.usn.mob3000.data.network.SupabaseClientWrapper
      */
     suspend fun insertFriendRequest(friendRequestItem: FriendRequestsDto): Result<Unit> = withContext(Dispatchers.IO) {
         try {
-            supabaseClient.from("friend_requests").insert(friendRequestItem)
+            supabaseClient
+                .from("friend_requests")
+                .insert(friendRequestItem)
+
             Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
+        } catch (error: Exception) {
+            Result.failure(error)
         }
     }
+
     /**
-     *
      * This function is used after a friend request is accepted, to join the two users in a row.
      *
      * @param friend The [FriendsDto] object containing the data for the new friend relationship.
@@ -120,9 +127,10 @@ import no.usn.mob3000.data.network.SupabaseClientWrapper
             supabaseClient
                 .from("friends")
                 .insert(friend)
+
             Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
+        } catch (error: Exception) {
+            Result.failure(error)
         }
     }
 }
