+++ app/src/main/java/no/usn/mob3000/domain/model/auth/UserStructs.kt
@@ -47,6 +47,7 @@ data class UserGameStats(
  * The user's profile struct.
  *
  * @author frigvid
+ * @contributor Husseinabdulameer11
  * @created 2024-10-22
  */
 data class UserProfile(
