+++ app/src/main/java/no/usn/mob3000/domain/model/social/FriendData.kt
@@ -3,9 +3,11 @@ package no.usn.mob3000.domain.model.social
 import kotlinx.datetime.Instant
 
 /**
+ * The pure kotlin data class representing a friendship.
+ *
  * @author Husseinabdulameer11
- * @created on 2024-11-05
- **/
+ * @created 2024-11-05
+ */
 data class FriendData(
     val friendshipId: String,
     val friendsSince: Instant,
