+++ app/src/main/java/no/usn/mob3000/domain/model/social/FriendRequestData.kt
@@ -2,11 +2,12 @@ package no.usn.mob3000.domain.model.social
 
 import kotlinx.datetime.Instant
 
-
 /**
+ * The pure Kotlin data class representing a friend request.
+ *
  * @author Husseinabdulameer11
- * created on 05.11.2024
- **/
+ * @created 2024-11-05
+ */
 data class FriendRequestData(
     val friendRequestId: String,
     val createdAt: Instant,
@@ -15,4 +16,3 @@ data class FriendRequestData(
     val accepted: Boolean,
     val displayName: String? = null
 )
-
