+++ app/src/main/java/no/usn/mob3000/ui/Navigation.kt
@@ -20,13 +20,13 @@ import no.usn.mob3000.domain.viewmodel.auth.ForgotPasswordViewModel
 import no.usn.mob3000.domain.viewmodel.auth.LoginViewModel
 import no.usn.mob3000.domain.viewmodel.auth.LogoutViewModel
 import no.usn.mob3000.domain.viewmodel.auth.RegistrationViewModel
-import no.usn.mob3000.domain.viewmodel.socials.ProfileViewModel
 import no.usn.mob3000.domain.viewmodel.content.provideDocumentationViewModel
 import no.usn.mob3000.domain.viewmodel.content.provideFAQViewModel
 import no.usn.mob3000.domain.viewmodel.content.provideNewsViewModel
 import no.usn.mob3000.domain.viewmodel.game.ChessBoardViewModel
 import no.usn.mob3000.domain.viewmodel.game.GroupsViewModel
 import no.usn.mob3000.domain.viewmodel.game.OpeningsViewModel
+import no.usn.mob3000.domain.viewmodel.social.ProfileViewModel
 import no.usn.mob3000.ui.components.base.Routes
 
 /**
@@ -89,8 +89,16 @@ val LocalNavController = compositionLocalOf<NavHostController> { error("No NavCo
  * @param viewModel The generic application ViewModel.
  * @param loginViewModel The login state ViewModel.
  * @param logoutViewModel The logout state ViewModel.
+ * @param changeEmailViewModel The E-mail change ViewModel.
+ * @param changePasswordViewModel The password change ViewModel.
  * @param registrationViewModel The registration state ViewModel.
+ * @param forgotPasswordViewModel The forgotten password ViewModel.
  * @param deleteAccountViewModel The account deletion state ViewModel.
+ * @param authenticationViewModel The authentication status state ViewModel.
+ * @param profileViewModel The user profile ViewModel.
+ * @param openingsViewModel The chess openings ViewModel.
+ * @param groupsViewModel The chess groups ViewModel.
+ * @param chessBoardViewModel The chess board's ViewModel.
  * @param navController The navigation controller.
  * @see Destination
  * @see Viewport
@@ -111,11 +119,11 @@ fun Navigation(
     forgotPasswordViewModel: ForgotPasswordViewModel = viewModel(),
     deleteAccountViewModel: DeleteAccountViewModel = viewModel(),
     authenticationViewModel: AuthenticationViewModel = viewModel(),
-    navController: NavHostController = rememberNavController(),
     profileViewModel: ProfileViewModel = viewModel(),
     openingsViewModel: OpeningsViewModel = viewModel(),
     groupsViewModel: GroupsViewModel = viewModel(),
-    chessBoardViewModel: ChessBoardViewModel = viewModel()
+    chessBoardViewModel: ChessBoardViewModel = viewModel(),
+    navController: NavHostController = rememberNavController()
 ) {
     CompositionLocalProvider(LocalNavController provides navController) {
         NavHost(
