+++ app/src/main/java/no/usn/mob3000/ui/screens/user/ProfileAddFriendsScreen.kt
@@ -2,12 +2,27 @@ package no.usn.mob3000.ui.screens.user
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -24,6 +39,10 @@ import no.usn.mob3000.ui.theme.DefaultButton
 /**
  * Screen for allowing users to search for other users and sending them a friend request.
  *
+ * @param fetchNonFriends Callback function to get all users who are not friends with the user.
+ * @param nonFriendState State for whether or not the users are friends.
+ * @param sendFriendRequest Callback function to send a friend request.
+ * @param onUserClick Callback function to insert a friend request into the database.
  * @author frigvid
  * @contributors 258030, Husseinabdulameer11
  * @created 2024-10-11
@@ -68,7 +87,8 @@ fun ProfileAddFriendsScreen(
                         user = user,
                         onClick = {
                             onUserClick(user.userId)
-                            sendFriendRequest() }
+                            sendFriendRequest()
+                        }
                     )
                 }
             }
@@ -76,8 +96,12 @@ fun ProfileAddFriendsScreen(
     }
 }
 
+/**
+ * @param user The user's [UserProfile].
+ * @param onClick Arbitrary unit.
+ */
 @Composable
-fun UserListItem(
+private fun UserListItem(
     user: UserProfile,
     onClick: () -> Unit
 ) {
